streamlit_app/requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=10.0.0
faiss-gpu>=1.7.2
numpy>=1.24.0
open-clip-torch>=2.20.0
sentencepiece>=0.1.99
timm>=0.9.0
